==> scripts/01_rank-chunk.ipynb <==
import os
import pandas as pd
import janitor
import numpy as np
pd.set_option('display.max_columns', 100)

DATAPATH = "../../adult-data"        

# (fold cell) Load and prep web browsing data
# output = df with visits to adult sites
# usecols_web_desktop = ["caseid", "category", "private_domain", "visit_duration", "visit_time_local"]
usecols_web_desktop = ["private_domain"]
usecols = ["private_domain"]

df = (pd.concat([
        # Get web_mobile
        pd.read_csv(
            os.path.join(DATAPATH, "output/realityMine_web_mobile_2022-06-01_2022-06-30.csv"),
            usecols=usecols_web_desktop,
            low_memory=False,
        ), 
        # Get web_desktop
        pd.read_csv(
            os.path.join(DATAPATH, "output/realityMine_web_desktop_2022-06-01_2022-06-30.csv"),
            usecols=usecols_web_desktop,
            low_memory=False,
        ),
        # Get web
        (pd.read_csv(
            os.path.join(DATAPATH, "output/realityMine_web_2022-06-01_2022-06-30.csv"),
            usecols=usecols,
            low_memory=False,
        )
        )
    ])
     )
print("output = df with web browses")
print(f"{len(df)=:,}")
df.head()

df_domains_ranked = (
    df.reset_index()
    .groupby("private_domain").size().reset_index()
    .sort_values(0, ascending=False, ignore_index=True)
    .rename_column(0, "tt_visits")
)
df_domains_ranked

def split_dataframe(dataframe: pd.DataFrame, n_chunks: int):
    """Split a dataframe into n chunks of approximately equal size."""
    
    # calculate the approximate chunk size
    chunk_size = len(dataframe) // n_chunks
    
    # split the dataframe into chunks
    chunks = [dataframe[i:i+chunk_size] for i in range(0, len(dataframe), chunk_size)]
    
    if len(chunks) > n_chunks:  # More chunks than specified, b/c of remainder rows
        # Concatenate the last two chunks
        chunks = chunks[:-2] + [pd.concat([chunks[-2], chunks[-1]])]
    return chunks

for ix, chunk in enumerate(split_dataframe(df_domains_ranked.select_columns("private_domain"), n_chunks=8)):
    (chunk
     .assign(return_code=np.nan)
     .to_csv(f"chunk{ix}.csv", index=False)
    )

!ls -lh
