==> scripts/02a_query_virustotal_proxy.nbconvert.ipynb <==
!pip install free-proxy

import pandas as pd
import requests
import time
from datetime import datetime
import json
import logging
import os

logging.basicConfig(filename='12_nbconvert_proxy.log', level=logging.INFO)

# Request rate	4 lookups / min
# Daily quota	500 lookups / day
# Monthly quota	15.5 K lookups / month

sleep = 200  # sleep 200 seconds after each call

headers = {
    "accept": "application/json",
    # https://www.virustotal.com/gui/user/lsys/apikey
}

from requests.exceptions import ProxyError, ReadTimeout, ConnectTimeout
from fp.fp import FreeProxy

def get_proxy():
    try:
        proxy = FreeProxy(rand=True).get()
        scheme_proxy_map = {
            'http': proxy,
            'https': proxy,
        }
        response = requests.get('https://ip.oxylabs.io/ip', proxies=scheme_proxy_map, timeout=10, verify=False)
        #print(response.text)
        if len(response.text.split()) == 1:
            return proxy
        else:
            return None
    #except (ProxyError, ReadTimeout, ConnectTimeout) as error:
    except Exception as error:
        print('Unable to connect to the proxy: ', error)
        return None


while True:
    print('Get proxy...')
    proxy = get_proxy()
    if proxy:
        print(proxy)
        break
    time.sleep(5)

scheme_proxy_map = {
    'http': proxy,
    'https': proxy,
}

chunk = "data/chunk2.csv"
df = pd.read_csv(chunk)

for ix, row in df.iterrows():
    now = datetime.now().strftime("%d %b %Y %H:%M:%S")
    domain = row["private_domain"]
    return_code = row["return_code"]
    
    ## Move on if already queried
    if return_code == 200:
        logging.info(f"({1+ix}/{len(df)}, {now}): Return code: {int(return_code)} - {domain}")
        continue

    if os.path.exists(f"payloads_json/{domain}.json"):
        logging.info(f"({1+ix}/{len(df)}, {now}): Return code: {return_code} - {domain}")
        continue

    while True:
        try:
            ## Query
            # https://developers.virustotal.com/reference/domain-info
            vt_url = f"https://www.virustotal.com/api/v3/domains/{domain}"
            response = requests.get(vt_url, headers=headers, proxies=scheme_proxy_map, timeout=15, verify=False)
            break
        except Exception as e:
            logging.error(f'Error: {e}')
            # Get new proxy
            while True:
                print('Get proxy...')
                proxy = get_proxy()
                if proxy:
                    print(proxy)
                    break
                time.sleep(5)
            scheme_proxy_map = {
                'http': proxy,
                'https': proxy,
            }
    ## Save payload
    return_code = response.status_code 
    if return_code == 200:
        with open(f"payloads_json/{domain}.json", "wb") as f:
            f.write(response.content)

    ## Update csv to indicate row is already queried
    df.at[ix, "return_code"] = return_code

    logging.info(f"({1+ix}/{len(df)}, {now}): Return code: {return_code} - {domain}")

    time.sleep(sleep)

df.to_csv(chunk, index=False)

df.to_csv(chunk, index=False)

df.query("return_code==200")
