==> scripts/06_ind_data.ipynb <==
import os
import pandas as pd
import janitor
from IPython.display import display
from inequality.gini import Gini
from scipy.stats import circmean
import warnings

warnings.filterwarnings("ignore")

import numpy as np

FP_IND_DOMAIN = "../data/yg_ind_domain.csv.gz"
FP_VISITS_DATA = "../data/6.3m-ind-domain-data.csv.gz"
FP_VT_LABELS = "../data/yg_virustotal_dat.csv"
FP_IND_DEMO = "../data/profile.csv"
FILEPATHS = [value for name, value in globals().items() if name.startswith("FP_")]
for file_path in FILEPATHS:
    try:
        assert os.path.exists(file_path), f"File not found: {file_path}"
    except AssertionError:
        print(f"File not found: {file_path}.")

df_visits = (
    pd.read_csv(FP_VISITS_DATA)
    # ==================================================
    # Merge to VT
    .merge(
        (pd.read_csv(FP_VT_LABELS)),
        how="left",
        left_on="private_domain",
        right_on="filename",
        validate="m:1",
    ).assign(
        malicious_bool=lambda df_: np.where(df_["malicious"] >= 2, True, False),
    )
    # ==================================================
    # time
    .assign(
        visit_time_local=lambda df_: pd.to_datetime(
            df_["visit_time_local"], format="ISO8601"
        ),
        hour_of_day=lambda df_: pd.to_datetime(df_["visit_time_local"]).dt.hour
        + pd.to_datetime(df_["visit_time_local"]).dt.minute / 60
        + pd.to_datetime(df_["visit_time_local"]).dt.second / 3600,
    )
)
display(df_visits.head())
df_visits.info()

df = (
    # =============================================================
    pd.read_csv(FP_IND_DOMAIN)
    .assign(
        duration_min=lambda df_: df_["duration"] / 60,
        duration_hr=lambda df_: df_["duration"] / 3600,
    )
    # =============================================================
    # Merge to VT
    .merge(
        (pd.read_csv(FP_VT_LABELS)),
        how="left",
        left_on="private_domain",
        right_on="filename",
        validate="m:1",
    )
    .assign(
        malicious_bool=lambda df_: np.where(df_["malicious"] >= 2, True, False),
        malicious_visits=lambda df_: df_["malicious_bool"] * df_["visits"],
        malicious_min=lambda df_: df_["malicious_bool"] * df_["duration_min"],
        malicious_hr=lambda df_: df_["malicious_bool"] * df_["duration_hr"],
        suspicious_bool=lambda df_: df_["suspicious"].astype("bool"),
    )
    # =============================================================
    # Aggregate to ind.
    .remove_columns(["private_domain", "filename", 'forcepoint', 'sophos', 'bitdefender', 'comodo', 'alphamnt'])
    .groupby("caseid")
    .sum()
    .reset_index()
    # =============================================================
    # Get singleton proportions
    # malicious
    .merge(
        (
            df_visits.query("malicious_bool")
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .assign(singleton_mal=lambda df_: np.where(df_["visits"] == 1, 1, 0))
            .groupby("caseid")["singleton_mal"]
            .mean()
            .reset_index()
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # non-malicious
    .merge(
        (
            df_visits.query("malicious==0")
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .assign(singleton_nonmal=lambda df_: np.where(df_["visits"] == 1, 1, 0))
            .groupby("caseid")["singleton_nonmal"]
            .mean()
            .reset_index()
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # All
    .merge(
        (
            df_visits
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .assign(singleton=lambda df_: np.where(df_["visits"] == 1, 1, 0))
            .groupby("caseid")["singleton"]
            .mean()
            .reset_index()
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # =============================================================
    # Get concentration measures
    # malicious
    .merge(
        (
            df_visits.query("malicious_bool")
            .reset_index(drop=True)
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .groupby("caseid")["visits"]
            .apply(lambda x: Gini(x.values).g)
            .reset_index(name="gini_mal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # non-malicious
    .merge(
        (
            df_visits.query("malicious==0")
            .query("suspicious==0")
            .reset_index(drop=True)
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .groupby("caseid")["visits"]
            .apply(lambda x: Gini(x.values).g)
            .reset_index(name="gini_nonmal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # All
    .merge(
        (
            df_visits
            .reset_index(drop=True)
            .groupby(["caseid", "private_domain"])
            .size()
            .reset_index()
            .rename_column(0, "visits")
            .groupby("caseid")["visits"]
            .apply(lambda x: Gini(x.values).g)
            .reset_index(name="gini")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )    
    # =============================================================
    # get unique sites visited
    # malicious
    .merge(
        (
            df_visits.query("malicious_bool")
            .groupby("caseid")["private_domain"]
            .nunique()
            .reset_index(name="n_uniques_mal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    .assign(n_uniques_mal=lambda df_: df_["n_uniques_mal"].fillna(0).astype(int))
    # non-malicious
    .merge(
        (
            df_visits.query("malicious==0")
            .query("suspicious==0")
            .groupby("caseid")["private_domain"]
            .nunique()
            .reset_index(name="n_uniques_nonmal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # All
    .merge(
        (
            df_visits
            .groupby("caseid")["private_domain"]
            .size()
            .reset_index(name="n_uniques")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )    
    # =============================================================
    # Average duration of visits
    # malicious
    .merge(
        (
            df_visits.query("malicious_bool")
            .groupby("caseid")["visit_duration"]
            .mean()
            .reset_index()
            .rename_column("visit_duration", "duration_mean_mal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )    
    # non-malicious
    .merge(
        (
            df_visits.query("malicious==0")
            .query("suspicious==0")
            .groupby("caseid")["visit_duration"]
            .mean()
            .reset_index()
            .rename_column("visit_duration", "duration_mean_nonmal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )  
    # All
    .merge(
        (
            df_visits
            .groupby("caseid")["visit_duration"]
            .mean()
            .reset_index()
            .rename_column("visit_duration", "duration_mean")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )     
    # =============================================================
    # Average time-of-day of visits
    # malicious
    .merge(
        (
            df_visits.query("malicious_bool")
            .groupby("caseid")["hour_of_day"]
            .apply(lambda x: circmean(x, high=24, low=0))
            .reset_index()
            .rename_column("hour_of_day", "tod_mean_mal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # non-malicious
    .merge(
        (
            df_visits.query("malicious==0")
            .query("suspicious==0")
            .groupby("caseid")["hour_of_day"]
            .apply(lambda x: circmean(x, high=24, low=0))
            .reset_index()
            .rename_column("hour_of_day", "tod_mean_nonmal")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )
    # All
    .merge(
        (
            df_visits
            .groupby("caseid")["hour_of_day"]
            .apply(lambda x: circmean(x, high=24, low=0))
            .reset_index()
            .rename_column("hour_of_day", "tod_mean")
        ),
        how="left",
        on="caseid",
        validate="1:1",
    )    
    # ==================================================
    # Merge to ind demo (n = 1200)
    .merge(
        (
            pd.read_csv(FP_IND_DEMO).assign(
                # https://github.com/themains/bad_domains/blob/main/data/codebook.pdf
                gender_lab=lambda df_: df_["gender"].replace({1: "Male", 2: "Female"}),
                race_lab=lambda df_: df_["race"].replace(
                    {
                        1: "White",
                        2: "Black",
                        3: "Hispanic",
                        4: "Asian",
                        5: "Other",
                        6: "Other",
                        7: "Other",
                        8: "Other",
                    }
                ),
                educ_lab=lambda df_: df_["educ"].replace(
                    {
                        1: "HS or Below",
                        2: "HS or Below",
                        3: "Some college",
                        4: "Some college",
                        5: "College",
                        6: "Postgrad",
                    }
                ),
                agegroup_lab=lambda df_: pd.cut(
                    df_["birthyr"],
                    # early baby boomers
                    # late baby boomers/ early genX
                    # genX, early millenials
                    # millenials
                    # genZ
                    bins=[1929, 1958, 1973, 1988, 1998, 2003],
                    labels=["65+", "50-64", "35-49", "25-34", "<25"],
                ),
            )
        ),
        how="left",
        on="caseid",
        validate="m:1",
    )
)
df.to_csv("../data/ind_data.csv", index=False)
display(df.head())
df.info(verbose=True)
