==> scripts/10_sample_comparison.ipynb <==
import pandas as pd
import janitor
import numpy as np
pd.set_option('display.max_columns', 100)
from utilities import pandas_to_tex
from IPython.display import display
import warnings
warnings.filterwarnings('ignore')

## CPS data

# https://cps.ipums.org/cps-action/variables/RACE#codes_section
race_codes_cps = {
    100: "White",
    200: "Black",
    300: "American Indian/Aluet/Eskimo",
    650: "Asian or Pacific Islander",
    651: "Asian only",
    652: "Hawaiian/Pacific Islander only",
    700: "Other (single) race, n.e.c.",
    801: "White-Black",
    802: "White-American Indian",
    803: "White-Asian",
    804: "White-Hawaiian/Pacific Islander",
    805: "Black-American Indian",
    806: "Black-Asian",
    807: "Black-Hawaiian/Pacific Islander",
    808: "American Indian-Asian",
    809: "Asian-Hawaiian/Pacific Islander",
    810: "White-Black-American Indian",
    811: "White-Black-Asian",
    812: "White-American Indian-Asian",
    813: "White-Asian-Hawaiian/Pacific Islander",
    814: "White-Black-American Indian-Asian",
    815: "American Indian-Hawaiian/Pacific Islander",
    816: "White-Black-Hawaiian/Pacific Islander",
    817: "White-American Indian-Hawaiian/Pacific Islander",
    818: "Black-American Indian-Asian",
    819: "White-American Indian-Asian-Hawaiian/Pacific Islander",
    820: "Two or three races, unspecified",
    830: "Four or five races, unspecified",
    999: "Blank",
}
race_map = {
    "White": "White",
    "Black": "Black",
    "American Indian/Aluet/Eskimo": "Other",
    "Asian or Pacific Islander": "Asian",
    "Asian only": "Asian",
    "Hawaiian/Pacific Islander only": "Asian",
    "Other (single) race, n.e.c.": "Other",
    "White-Black": "Other",
    "White-American Indian": "Other",
    "White-Asian": "Other",
    "White-Hawaiian/Pacific Islander": "Other",
    "Black-American Indian": "Other",
    "Black-Asian": "Other",
    "Black-Hawaiian/Pacific Islander": "Other",
    "American Indian-Asian": "Other",
    "Asian-Hawaiian/Pacific Islander": "Asian",
    "White-Black-American Indian": "Other",
    "White-Black-Asian": "Other",
    "White-American Indian-Asian": "Other",
    "White-Asian-Hawaiian/Pacific Islander": "Other",
    "White-Black-American Indian-Asian": "Other",
    "American Indian-Hawaiian/Pacific Islander": "Other",
    "White-Black-Hawaiian/Pacific Islander": "Other",
    "White-American Indian-Hawaiian/Pacific Islander": "Other",
    "Black-American Indian-Asian": "Other",
    "White-American Indian-Asian-Hawaiian/Pacific Islander": "Other",
    "Two or three races, unspecified": "Other",
    "Four or five races, unspecified": "Other",
    "Blank": np.nan,
}

# https://cps.ipums.org/cps-action/variables/HISPAN#codes_section
hispanic_codes_cps = {
    0: "Not Hispanic",
    100: "Mexican",
    102: "Mexican American",
    103: "Mexicano/Mexicana",
    104: "Chicano/Chicana",
    108: "Mexican (Mexicano)",
    109: "Mexicano/Chicano",
    200: "Puerto Rican",
    300: "Cuban",
    400: "Dominican",
    500: "Salvadoran",
    600: "Other Hispanic",
    610: "Central/South American",
    611: "Central American (excluding Salvadoran)",
    612: "South American",
    901: "Do not know",
    902: "Not available/no response"
}

hispanic_mapping = {
    0: "Non-Hispanic",  # Not Hispanic
    100: "Hispanic",
    102: "Hispanic",
    103: "Hispanic",
    104: "Hispanic",
    108: "Hispanic",
    109: "Hispanic",
    200: "Hispanic",
    300: "Hispanic",
    400: "Hispanic",
    500: "Hispanic",
    600: "Hispanic",
    610: "Hispanic",
    611: "Hispanic",
    612: "Hispanic",
    901: None,  # Do not know
    902: None   # Not available/no response
}


# https://cps.ipums.org/cps-action/variables/EDUC#codes_section
education_mapping = {
    # HS or Below
    0: "HS or Below", 1: "HS or Below", 2: "HS or Below", 
    10: "HS or Below", 11: "HS or Below", 12: "HS or Below",
    13: "HS or Below", 14: "HS or Below", 20: "HS or Below",
    21: "HS or Below", 22: "HS or Below", 30: "HS or Below",
    31: "HS or Below", 32: "HS or Below", 40: "HS or Below",
    50: "HS or Below", 60: "HS or Below", 70: "HS or Below",
    71: "HS or Below", 72: "HS or Below", 73: "HS or Below",
    # Some college
    80: "Some college", 81: "Some college", 
    90: "Some college", 91: "Some college", 92: "Some college",
    # College
    100: "College", 110: "College", 111: "College",
    # Postgrad
    120: "Postgrad", 121: "Postgrad", 122: "Postgrad", 
    123: "Postgrad", 124: "Postgrad", 125: "Postgrad",
    # Missing
    999: None,
}


age_bins = [-1, 17, 25, 34, 49, 64, float('inf')]
age_labels = ["<18", "18--24", "25--34", "35--49", "50--64", "65+"]

region_codes_cps = {
    11: "Northeast",
    12: "Northeast",
    21: "Midwest",
    22: "Midwest",
    31: "South",
    32: "South",
    33: "South",
    41: "West",
    42: "West",
    97: "Unknown"
}

df_cps = (
    pd.read_csv("../data/cps_00002.csv.gz")
    .clean_names()
    .assign(
        race_lab_cps=lambda df_: df_["race"].replace(race_codes_cps),
        race_lab_nohisp=lambda df_: df_["race_lab_cps"].replace(race_map),
        hispanic_lab=lambda df_: df_["hispan"].replace(hispanic_codes_cps),
        race_lab=lambda df_: np.where(df_["hispanic_lab"]=="Not Hispanic", df_["race_lab_nohisp"], "Hispanic"),
        educ_lab=lambda df_: df_["educ"].replace(education_mapping),
        agegroup_lab=lambda df_: pd.cut(df_['age'], bins=age_bins, labels=age_labels),
        # https://cps.ipums.org/cps-action/variables/SEX#codes_section
        gender_lab=lambda df_: df_["sex"].map({1: "Male", 2: "Female"}),
        # https://cps.ipums.org/cps-action/variables/VOTED#codes_section
        voted_lab=lambda df_: np.where(df_["voted"]==2, "Voted", "Did not vote/NA"),
#         voted_lab=np.nan,
        region_lab=lambda df_: df_["region"].replace(region_codes_cps),
    )
    .query("agegroup_lab!='<18'")
)
display(df_cps.head())
df_cps.info()

## YouGov

region_codes_yougov = {
    1: "Northeast",
    2: "Midwest",
    3: "South",
    4: "West",
    -1: np.nan
}
voting_mapping_yougov = {
    -1: "Did not vote/NA",  # No Data
    1: "Voted",       # Joe Biden
    2: "Voted",       # Donald Trump
    3: "Voted",       # Jo Jorgensen
    4: "Voted",       # Howie Hawkins
    5: "Voted",       # Other
    6: "Did not vote/NA"    # Did not vote for President
}

agegroup_yg = {
    "65+": "65+",
    "50--65": "50--64",
    "35--50": "35--49",
    "25--35": "25--34",
    "<25": "18--24",
}
df_yougov = (
    pd.read_csv("../data/ind_data.csv")
    .assign(
        region_lab=lambda df_: df_["region"].replace(region_codes_yougov),
        voted_lab=lambda df_: df_["presvote20post"].map(voting_mapping_yougov),
        agegroup_lab=lambda df_: df_["agegroup_lab"].map(agegroup_yg),
        age=lambda df_: 2022-df_["birthyr"],
    )
)
df_yougov.head()

## Compare

# import pandas as pd

# # Define a function to summarize field distributions
# def summarize_sample(df, fields):
#     summary = {}
#     for field in fields:
#         summary[field] = df[field].value_counts(normalize=True).reset_index()
#         summary[field].columns = [field, 'Proportion']
#     return summary

# # Fields to compare
# fields_to_compare = ['gender_lab', 'race_lab', 'educ_lab', 'agegroup_lab', 'region_lab']

# # Summarize both datasets
# summary_cps = summarize_sample(df_cps, fields_to_compare)
# summary_yougov = summarize_sample(df_yougov, fields_to_compare)

# # Combine summaries for comparison
# comparison = {}
# for field in fields_to_compare:
#     comparison[field] = pd.merge(
#         summary_cps[field],
#         summary_yougov[field],
#         on=field,
#         how='outer',
#         suffixes=('_CPS', '_YouGov')
#     ).fillna(0)

# # Display the comparison tables for each field
# for field, table in comparison.items():
#     print(f"--- {field} ---")
#     print(table)
#     print()


# var = "gender_lab"

# _dfyg = (
#     df_yougov[var]
#     .value_counts(normalize=True)
#     .reset_index()
#     .rename_column(var, "yougov")
# )
# _dfcps = (
#     df_cps[var]
#     .value_counts(normalize=True)
#     .reset_index()
#     .rename_column(var, "cps")
# )
# _dfsumm = (
#     _dfyg
#     .merge(_dfcps, on="index", validate="1:1")
# )
# _dfsumm

catvars = ["gender_lab", "race_lab", "educ_lab", "agegroup_lab", "region_lab",]
agegroup_order = ["<18", "18--24", "25--34", "35--49", "50--64", "65+"]
region_order = ["West", "Midwest", "Northeast", "South"]

for var in catvars:
    _dfyg = (
        df_yougov[var]
        .value_counts(normalize=True)
        .reset_index()
        .rename_column(var, "yougov")
    )
    _dfcps = (
        df_cps[var]
        .value_counts(normalize=True)
        .reset_index()
        .rename_column(var, "cps")
    )
    _dfsumm = (
        _dfyg
        .merge(_dfcps, on="index", validate="1:1")
        .round(3)
    )
    
    # reorder agegroup
    if var == "agegroup_lab":
        _dfsumm["index"] = (
            pd.Categorical(
                _dfsumm["index"], categories=agegroup_order, ordered=True)
        )
        _dfsumm = _dfsumm.sort_values("index")
    
    # reorder regions    
    if var == "region_lab":
        _dfsumm["index"] = pd.Categorical(
            _dfsumm["index"], categories=region_order, ordered=True
        )
        _dfsumm = _dfsumm.sort_values("index")
        _dfsumm["index"] = _dfsumm["index"].apply(lambda x: x + " region")
        
    if var=="educ_lab":
#         _dfsumm["ed"]
        _dfsumm.loc[_dfsumm['index'] == 'Postgrad', 'index'] = 'Postgraduate degree'
        _dfsumm.loc[_dfsumm['index'] == 'HS or Below', 'index'] = 'High school or below'
        _dfsumm.loc[_dfsumm['index'] == 'College', 'index'] = 'College degree'

    pandas_to_tex(_dfsumm, f"../tabs/yg_cps_{var}")
    display(_dfsumm)

var = "age"
mean_yougov = df_yougov[var].mean()
mean_cps = df_cps[var].mean()

_dfsumm = (
    pd.DataFrame({"var": ["Age (mean)"], "yougov": mean_yougov, "cps": mean_cps})
    .round(1)
)
pandas_to_tex(_dfsumm, "../tabs/yg_cps_agemean")
_dfsumm

## Nationscape

# https://www.voterstudygroup.org/downloads/nationscape?key=1015588
race_mapping_ns = {
    "Black, or African American": "Black",
    "White": "White",
    "Asian (Filipino)": "Asian",
    "Asian (Vietnamese)": "Asian",
    "Asian (Asian Indian)": "Asian",
    "Asian (Korean)": "Asian",
    "Pacific Islander (Other)": "Other",
    "Pacific Islander (Native Hawaiian)": "Other",
    "Pacific Islander (Guamanian)": "Other",
    "American Indian or Alaska Native": "Other",
    "Some other race": "Other",
    "Asian (Chinese)": "Asian",
    "Asian (Japanese)": "Asian"
}
education_mapping_ns = {
    "3rd Grade or less": "HS or Below",
    "Middle School - Grades 4 - 8": "HS or Below",
    "Completed some high school": "HS or Below",
    "High school graduate": "HS or Below",
    "Completed some college, but no degree": "Some college",
    "Associate Degree": "Some college",
    "Other post high school vocational training": "Some college",
    "College Degree (such as B.A., B.S.)": "College",
    "Completed some graduate, but no degree": "Postgrad",
    "Masters degree": "Postgrad",
    "Doctorate degree": "Postgrad"
}

vote_2020_mapping_ns = {
    "Joe Biden": "Voted",
    "Donald Trump": "Voted",
    "Someone else": "Voted",
    "I abstained": "Did not vote/NA",
    "I don't recall": "Did not vote/NA",
    "Not Asked": "Did not vote/NA",
    None: "Did not vote/NA"  # Handle NaN or missing values
}

df_ns = (
    pd.read_stata("../data/Nationscape-Weekly-Materials-DTA-2021Dec/phase_3_v20210301/ns20210112/ns20210112.dta")
    .assign(
        race_lab_nohisp=lambda df_: df_["race_ethnicity"].map(race_mapping_ns),
        race_lab=lambda df_: np.where(df_["hispanic"]=="Not Hispanic", df_["race_lab_nohisp"], "Hispanic"),
        gender_lab=lambda df_: df_["gender"],
        educ_lab=lambda df_: df_["education"].map(education_mapping_ns),
        agegroup_lab=lambda df_: pd.cut(df_['age'], bins=age_bins, labels=age_labels),
        region_lab=lambda df_: df_["census_region"],
        voted_lab=lambda df_: df_["vote_2020_retro"].map(vote_2020_mapping_ns),
    )
)
display(df_ns.head())
df_ns.info(verbose=True)

catvars = ["gender_lab", "race_lab", "educ_lab", "agegroup_lab", "region_lab", "voted_lab"]

for var in catvars:
    _dfyg = (
        df_yougov[var]
        .value_counts(normalize=True)
        .reset_index()
        .rename_column(var, "yougov")
    )
    _dfcps = (
        df_cps[var]
        .value_counts(normalize=True)
        .reset_index()
        .rename_column(var, "cps")
    )
    _dfns = (
        df_ns[var]
        .value_counts(normalize=True)
        .reset_index()
        .rename_column(var, "nscape")
    )    
    _dfsumm = (
        _dfyg
        .merge(_dfcps, on="index", validate="1:1")
        .merge(_dfns, on="index", validate="1:1")
        .round(3)
    )
    
    # reorder agegroup
    if var == "agegroup_lab":
        _dfsumm["index"] = (
            pd.Categorical(
                _dfsumm["index"], categories=agegroup_order, ordered=True)
        )
        _dfsumm = _dfsumm.sort_values("index")
    
    # reorder regions    
    if var == "region_lab":
        _dfsumm["index"] = pd.Categorical(
            _dfsumm["index"], categories=region_order, ordered=True
        )
        _dfsumm = _dfsumm.sort_values("index")
        _dfsumm["index"] = _dfsumm["index"].apply(lambda x: x + " region")
        
    if var=="educ_lab":
#         _dfsumm["ed"]
        _dfsumm.loc[_dfsumm['index'] == 'Postgrad', 'index'] = 'Postgraduate degree'
        _dfsumm.loc[_dfsumm['index'] == 'HS or Below', 'index'] = 'High school or below'
        _dfsumm.loc[_dfsumm['index'] == 'College', 'index'] = 'College degree'

    if var=="voted_lab":
        _dfsumm['cps'] = "---"

#     pandas_to_tex(_dfsumm, f"../tabs/yg_cps_nscape_{var}")
    display(_dfsumm)
